[Feat] MainNavigator 추가, AppManager 적용
